chore: remove unused code and add Process.Kill()

+++ Editor/ProcessUtils.cs
@@ -2,63 +2,35 @@ using System;
 using System.Diagnostics;
 using System.Threading.Tasks;
 using System.Threading;
+using Debug = UnityEngine.Debug;
 
 namespace Neovim.Editor
 {
   public static class ProcessUtils
   {
-    public static bool RunProcess(string app, string args, ProcessWindowStyle winStyle, bool createNoWindow = false,
-    bool useShellExecute = false)
+    /// this is mainly used to avoid cross-platform issues where, for instance, on Windows
+    /// a call to WaitForExit(timeout) may not behave in an expected way.
+    public static void RunProcessAndKillAfter(string app, string args, int timeout = 200)
     {
-      bool success = false;
-      try
+      using (Process p = new())
       {
-        using (Process p = new())
+        p.StartInfo.FileName = app;
+        p.StartInfo.Arguments = args;
+        p.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
+        p.StartInfo.CreateNoWindow = true;
+        p.StartInfo.UseShellExecute = false;
+
+        p.Start();
+
+        if (!p.WaitForExit(timeout))
         {
-          p.StartInfo.FileName = app;
-          p.StartInfo.Arguments = args;
-          p.StartInfo.WindowStyle = winStyle;
-          p.StartInfo.CreateNoWindow = false;
-          p.StartInfo.UseShellExecute = useShellExecute;
-
-          p.Start();
-
-          success = true;
-        }
-      }
-      catch (Exception) { }
-      return success;
-    }
-
-
-    /// runs process (usually some shell cmd) and exits (i.e., gets killed) almost immediately
-    /// (after at most 50ms) this is mainly used to avoid cross-platform issues where, for
-    /// instance, on Windows a call to WaitForExit(timeout) may not behave in an expected way.
-    public static void RunProcessAndExitImmediately(string app, string args)
-    {
-      try
-      {
-        using (Process p = new())
-        {
-          p.StartInfo.FileName = app;
-          p.StartInfo.Arguments = args;
-          p.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
-          p.StartInfo.CreateNoWindow = true;
-          p.StartInfo.UseShellExecute = false;
-
-          p.Start();
-
-          // WaitForExitAsync is only available in newer .NET versions
-          p.WaitForExit(50);
-
           p.Kill();
         }
       }
-      catch (Exception) { }
     }
 
 
-    public static bool RunShellCmd(string cmd, int timeout = 500)
+    public static bool RunShellCmd(string cmd, int timeout = 200)
     {
       bool success = false;
       try
@@ -66,9 +38,9 @@ namespace Neovim.Editor
         using (Process p = new())
         {
 #if UNITY_EDITOR_LINUX
-            string escapedArgs = escapedArgs = cmd.Replace("\"", "\\\"");
-            p.StartInfo.FileName = Environment.GetEnvironmentVariable("SHELL");
-            p.StartInfo.Arguments = $"-c \"{escapedArgs}\"";
+          string escapedArgs = escapedArgs = cmd.Replace("\"", "\\\"");
+          p.StartInfo.FileName = Environment.GetEnvironmentVariable("SHELL");
+          p.StartInfo.Arguments = $"-c \"{escapedArgs}\"";
 #else // UNITY_EDITOR_WIN
           p.StartInfo.FileName = "cmd.exe";
           p.StartInfo.Arguments = $"/C \"{cmd}\"";
@@ -76,15 +48,20 @@ namespace Neovim.Editor
           p.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
           p.StartInfo.CreateNoWindow = true;
           p.StartInfo.UseShellExecute = false;
+
           p.Start();
-          p.WaitForExit();
-          success = p.ExitCode == 0;
+
+          if (p.WaitForExit(timeout))
+          {
+            success = p.ExitCode == 0;
+          }
+          else
+          {
+            p.Kill();
+          }
         }
       }
-      catch (Exception)
-      {
-        success = false;
-      }
+      catch (Exception) { }
       return success;
     }
 
@@ -107,21 +84,18 @@ namespace Neovim.Editor
           p.StartInfo.CreateNoWindow = true;
           p.StartInfo.UseShellExecute = false;
           p.Start();
-          if (!p.WaitForExit(timeout))
-          {
-            success = false;
-          }
-          else
+          if (p.WaitForExit(timeout))
           {
             // which/where.exe returns 0 if the supplied cmd was found
             success = p.ExitCode == 0;
           }
+          else
+          {
+            p.Kill();
+          }
         }
       }
-      catch (Exception)
-      {
-        success = false;
-      }
+      catch (Exception) { }
       return success;
     }
   }
